Added uninstall options. Abstracted install/uninstall functions to separate script
